Fix shell_out "No such file or directory" issue
A string passed to `shell_out` method as a command argument that contains binary and arguments, additionally escaped in underlayed methods and interpreted as a single binary to execute. This leads to errors like this:

```
================================================================================
Error executing action `start` on resource 'runit_service[myservice01]'
================================================================================

Errno::ENOENT
-------------
No such file or directory - /usr/bin/sv status /etc/service/myservice01
```

In this commit it is suggested to pass a binary and arguments as an array. This allows properly escape each element of the whole command in underlayed methods.